Добавит модуляцию
