+++ js/messages.js
@@ -41,9 +41,9 @@ const showErrorMessage = () => {
 };
 
 function hideMessage() {
-  const messageElemnt =
+  const messageElement =
     document.querySelector('.success') || document.querySelector('.error');
-  messageElemnt.remove();
+  messageElement.remove();
   document.removeEventListener('keydown', onMessageEscKeydown);
   document.removeEventListener('click', onOverlayClick);
   bodyElement.style.overflow = 'auto';
